refactor: retain project-list-view filename
