+++ src/components/project-list-view.tsx
